+++ nhaystack/__init__.py
@@ -2,4 +2,6 @@
 
 __title__ = 'nhaystack'
 __author__ = 'Noveo Group'
-__version__ = '0.0.1'
+
+VERSION = (0, 0, 2)  # PEP 386
+__version__ = '.'.join(str(x) for x in VERSION)