+++ tests/conftest.py
@@ -1,0 +1,9 @@
+# coding: utf-8
+
+from __future__ import absolute_import
+
+import os
+
+
+def pytest_configure():
+    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'tests.settings')